feat: moved LoadingErrorMessage to shared folder
