+++ src/Analysis/GWASResults/Views/Home/HomeTable/HomeTable.jsx
@@ -16,7 +16,7 @@ import {
 } from './tableDataProcessing/tableDataProcessing';
 import VIEWS from '../../../Utils/ViewsEnumeration';
 import isIterable from '../../../Utils/isIterable';
-import LoadingErrorMessage from '../../../Components/LoadingErrorMessage/LoadingErrorMessage';
+import LoadingErrorMessage from '../../../../SharedUtils/LoadingErrorMessage/LoadingErrorMessage';
 
 import './HomeTable.css';
