+++ src/Analysis/GWASResults/Views/Results/ResultsPheWeb/QQPlotModal/QQPlotModal.jsx
@@ -6,7 +6,7 @@ import QQPlot from '../../../../Components/Diagrams/QQPlot/QQPlot';
 import SharedContext from '../../../../Utils/SharedContext';
 import { getDataForWorkflowArtifact } from '../../../../Utils/gwasWorkflowApi';
 import queryConfig from '../../../../../SharedUtils/QueryConfig';
-import LoadingErrorMessage from '../../../../Components/LoadingErrorMessage/LoadingErrorMessage';
+import LoadingErrorMessage from '../../../../../SharedUtils/LoadingErrorMessage/LoadingErrorMessage';
 
 const QQPlotModal = ({ modalOpen, setModalOpen }) => {
   const { selectedRowData } = useContext(SharedContext);
