+++ src/Analysis/GWASResults/Views/Results/ResultsPheWeb/ResultsPheWeb.jsx
@@ -5,7 +5,7 @@ import * as d3 from 'd3-selection';
 import SharedContext from '../../../Utils/SharedContext';
 import { getDataForWorkflowArtifact } from '../../../Utils/gwasWorkflowApi';
 import queryConfig from '../../../../SharedUtils/QueryConfig';
-import LoadingErrorMessage from '../../../Components/LoadingErrorMessage/LoadingErrorMessage';
+import LoadingErrorMessage from '../../../../SharedUtils/LoadingErrorMessage/LoadingErrorMessage';
 import QQPlotModal from './QQPlotModal/QQPlotModal';
 import ManhattanPlot from '../../../Components/Diagrams/ManhattanPlot/ManhattanPlot';
 import TopLociTable from './TopLociTable/TopLociTable';
